+++ source/spark_sql_migrations/spark_sql_migrations/constants/migrations_constants.py
@@ -1,12 +1,3 @@
-class SchemaMigrationConstants:
-    default_location = "schema-migration"
-    schema_name = "schema_migration"
-    table_name = "executed_migrations"
-    migration_scripts_folder_path = "package.schema_migration.migration_scripts"
-    current_state_schemas_folder_path = "package.schema_migration.current_state_scripts.schemas"
-    current_state_tables_folder_path = "package.schema_migration.current_state_scripts.tables"
-
-
 class ColNames:
     migration_name = "migration_name"
     execution_datetime = "execution_datetime"
